# contracts/openapi/freshflow.yaml
openapi: 3.1.0
info:
  title: FreshFlow API
  version: 0.4.0
  description: Catalog, cart, idempotent checkout, delivery and ClickHouse analytics API exposed by api-gateway.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/catalog/products:
    get:
      operationId: listProducts
      tags: [Catalog]
      summary: List active products and current availability
      responses:
        "200":
          description: Product catalog
          content:
            application/json:
              schema:
                type: object
                required: [products]
                properties:
                  products:
                    type: array
                    items: {$ref: "#/components/schemas/Product"}
        "503": {$ref: "#/components/responses/ServiceUnavailable"}
...
components:
...
  responses:
...
    ServiceUnavailable:
      description: A required dependency is unavailable
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  schemas:
...
    Product:
      type: object
      required: [id, sku, name, description, price_minor, currency, available_quantity]
      properties:
        id: {type: string, format: uuid}
        sku: {type: string}
        name: {type: string}
        description: {type: string}
        price_minor: {type: integer, minimum: 0}
        currency: {type: string, minLength: 3, maxLength: 3}
        available_quantity: {type: integer, minimum: 0}
...
    APIError:
      type: object
      required: [code, message, correlation_id]
      properties:
        code: {type: string}
        message: {type: string}
        correlation_id: {type: string}
        details: {}
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: {$ref: "#/components/schemas/APIError"}